separate out RedshiftColumn to its own module
