+++ dbt/adapters/redshift/column.py
@@ -1,0 +1,5 @@
+from dbt.adapters.base import Column
+
+
+class RedshiftColumn(Column):
+    pass  # redshift does not inherit from postgres here
