+++ dbt/adapters/redshift/impl.py
@@ -5,7 +5,7 @@ from dbt.adapters.sql import SQLAdapter
 from dbt.adapters.base.meta import available
 from dbt.adapters.postgres import PostgresAdapter
 from dbt.adapters.redshift import RedshiftConnectionManager
-from dbt.adapters.redshift import RedshiftColumn
+from dbt.adapters.redshift.column import RedshiftColumn
 from dbt.adapters.redshift import RedshiftRelation
 from dbt.events import AdapterLogger
 import dbt.exceptions
